Convert project to maven
