--- plugin.yml
@@ -1,20 +1,0 @@
-main: de.jeffclan.JeffChestSort.JeffChestSortPlugin
-name: ChestSort
-version: 3.5
-api-version: 1.13
-description: Allows automatic chest sorting
-author: mfnalex
-website: https://www.spigotmc.org/resources/1-13-chestsort.59773/
-prefix: ChestSort
-database: false
-loadbefore:
-- InvUnload
-commands:
-  chestsort:
-    description: Toggle automatic chest sorting
-    usage: /<command>
-    aliases: sort
-    permission: chestsort.use
-permissions:
-  chestsort.use:
-    description: Allows usage of automatic chest sorting
