--- src/de/jeffclan/utils/categories/910-tools.default.txt
@@ -1,10 +1,0 @@
-*_pickaxe
-*_shovel
-*_hoe
-*_axe
-flint_and_steel
-fishing_rod
-compass
-clock
-shears
-lead
