--- src/de/jeffclan/utils/categories/940-food.default.txt
@@ -1,30 +1,0 @@
-apple
-baked_potato
-beef
-beetroot
-beetroot_soup
-bread
-mushroom_stew
-porkchop
-cooked_porkchop
-cod
-salmon
-cooked_cod
-cooked_salmon
-cake
-cookie
-melon_slice
-dried_kelp
-cooked_beef
-chicken
-cooked_chicken
-carrot
-potato
-pumpkin_pie
-rabbit
-cooked_rabbit
-rabbit_stew
-mutton
-cooked_mutton
-wheat
-*_seeds
