--- src/de/jeffclan/utils/categories/950-redstone.default.txt
@@ -1,23 +1,0 @@
-dispenser
-note_block
-sticky_piston
-piston
-tnt
-lever
-*_pressure_plate
-redstone
-*_button
-redstone_lamp
-redstone_torch
-tripwire_hook
-trapped_chest
-daylight_detector
-redstone_block
-redstone_ore
-hopper
-dropper
-iron_trapdoor
-observer
-iron_door
-repeater
-comparator
