+++ src/main/java/de/jeffclan/JeffChestSort/JeffChestSortPlayerSetting.java
@@ -1,0 +1,15 @@
+package de.jeffclan.JeffChestSort;
+
+public class JeffChestSortPlayerSetting {
+
+	// Sorting enabled for this player?
+	boolean sortingEnabled;
+
+	// Did we already show the message how to activate sorting?
+	boolean hasSeenMessage = false;
+
+	JeffChestSortPlayerSetting(boolean sortingEnabled) {
+		this.sortingEnabled = sortingEnabled;
+	}
+
+}
