+++ src/main/resources/categories/900-valuables.default.txt
@@ -1,0 +1,27 @@
+diamond
+emerald
+diamond_ore
+emerald_ore
+iron_ingot
+iron_ore
+gold_ingot
+gold_ore
+gold_block
+diamond_block
+iron_block
+lapis_lazuli
+lapis_block
+obsidian
+beacon
+conduit
+scute
+elytra
+ender_pearl
+ender_eye
+*_horse_armor
+music_disc_*
+heart_of_the_sea
+nautilus_shell
+gold_nugget
+iron_nugget
+enchanted_book
