+++ src/main/resources/categories/960-wood.default.txt
