+++ src/main/resources/categories/970-stone.default.txt
