+++ src/main/resources/categories/980-plants.default.txt
